--- base/shared/src/main/scala/scalaz/ct/FoldableInstances.scala
@@ -1,4 +1,0 @@
-package scalaz
-package ct
-
-trait FoldableInstances {}
