--- base/shared/src/main/scala/scalaz/ct/FunctorFunctions.scala
@@ -1,6 +1,0 @@
-package scalaz
-package ct
-
-trait FunctorFunctions {
-  def map[F[_], A, B](fa: F[A])(f: A => B)(implicit F: Functor[F]): F[B] = F.map(fa)(f)
-}
