--- base/shared/src/main/scala/scalaz/ct/InvariantFunctorClass.scala
@@ -1,6 +1,0 @@
-package scalaz
-package ct
-
-trait InvariantFunctorClass[F[_]] {
-  def imap[A, B](ma: F[A])(f: A => B)(g: B => A): F[B]
-}
