--- base/shared/src/main/scala/scalaz/ct/PhantomSyntax.scala
@@ -1,10 +1,0 @@
-package scalaz
-package ct
-
-import scala.language.experimental.macros
-
-trait PhantomSyntax {
-  implicit final class ToPhantomOps[F[_], A](self: F[A]) {
-    def pmap[B](implicit ev: Phantom[F]): F[B] = macro meta.Ops.i_0
-  }
-}
