--- base/shared/src/main/scala/scalaz/ct/ProfunctorInstances.scala
@@ -1,4 +1,0 @@
-package scalaz
-package ct
-
-trait ProfunctorInstances {}
