--- base/shared/src/main/scala/scalaz/ct/TraversableFunctions.scala
@@ -1,7 +1,0 @@
-package scalaz
-package ct
-
-trait TraversableFunctions {
-  def sequence[T[_], F[_], A](tfa: T[F[A]])(implicit F: Applicative[F], T: Traversable[T]): F[T[A]] =
-    T.sequence(tfa)
-}
