+++ base/shared/src/main/scala/scalaz/ct/contravariant.scala
