+++ base/shared/src/main/scala/scalaz/ct/foldable.scala
@@ -1,6 +1,8 @@
 package scalaz
 package ct
 
+import scala.language.experimental.macros
+
 trait FoldableClass[F[_]] {
 
   def foldMap[A, B: Monoid](fa: F[A])(f: A => B): B
@@ -30,4 +32,15 @@ object FoldableClass {
 
   trait Alt[D <: Alt[D]]
 
+}
+
+trait FoldableInstances {}
+
+trait FoldableSyntax {
+  implicit final class ToFoldableOps[F[_], A](self: F[A]) {
+    def foldLeft[B](f: B)(g: (B, A) => B)(implicit ev: Foldable[F]): B = macro meta.Ops.ia_1_1
+    def foldRight[B](f: => B)(g: (A, => B) => B)(implicit ev: Foldable[F]): B = macro meta.Ops.ia_1_1
+    def foldMap[B](f: A => B)(implicit g: Monoid[B], ev: Foldable[F]): B = macro meta.Ops.i_1_1i
+    def toList(implicit ev: Foldable[F]): List[A] = macro meta.Ops.i_0
+  }
 }
