+++ base/shared/src/main/scala/scalaz/ct/monad.scala
@@ -1,6 +1,17 @@
 package scalaz
 package ct
 
+trait MonadClass[M[_]] extends ApplicativeClass[M] with BindClass[M]
+
+object MonadClass {
+
+  trait DeriveMap[M[_]] extends MonadClass[M] with BindClass.Alt[BindClass.DeriveFlatten[M]] {
+    final override def map[A, B](ma: M[A])(f: (A) => B): M[B] = flatMap(ma)(a => pure(f(a)))
+  }
+
+  trait Alt[D <: Alt[D]]
+}
+
 trait MonadInstances {
   implicit val optionMonad: Monad[Option] = instanceOf(new MonadClass[Option] with BindClass.DeriveFlatten[Option] {
     override def ap[A, B](oa: Option[A])(f: Option[A => B]): Option[B]      = oa.flatMap(a => f.map(_(a)))
