+++ base/shared/src/main/scala/scalaz/ct/phantom.scala
@@ -1,6 +1,10 @@
 package scalaz
 package ct
 
+import scalaz.data.Const
+
+import scala.language.experimental.macros
+
 trait PhantomClass[F[_]] extends FunctorClass[F] with ContravariantClass[F] {
   def pmap[A, B](ma: F[A]): F[B]
 }
@@ -17,4 +21,22 @@ object PhantomClass {
   }
 
   trait Alt[D <: Alt[D]]
+}
+
+trait PhantomFunctions {
+  def pmap[F[_], A, B](fa: F[A])(implicit F: Phantom[F]): F[B] = F.pmap(fa)
+}
+
+trait PhantomInstances {
+
+  implicit def const[R]: Phantom[Const[R, ?]] =
+    instanceOf(new PhantomClass.DeriveMapContramap[Const[R, ?]] {
+      def pmap[A, B](ma: Const[R, A]): Const[R, B] = ma.retag[B]
+    })
+}
+
+trait PhantomSyntax {
+  implicit final class ToPhantomOps[F[_], A](self: F[A]) {
+    def pmap[B](implicit ev: Phantom[F]): F[B] = macro meta.Ops.i_0
+  }
 }
