+++ base/shared/src/main/scala/scalaz/ct/profunctor.scala
@@ -1,6 +1,8 @@
 package scalaz
 package ct
 
+import scala.language.experimental.macros
+
 trait ProfunctorClass[F[_, _]] {
   def lmap[A, B, C](fab: F[A, B])(ca: C => A): F[C, B]
   def rmap[A, B, C](fab: F[A, B])(bc: B => C): F[A, C]
@@ -19,4 +21,14 @@ object ProfunctorClass {
   }
 
   trait Alt[D <: Alt[D]]
+}
+
+trait ProfunctorInstances {}
+
+trait ProfunctorSyntax {
+  implicit class ToProfunctorOps[F[_, _]: Profunctor, A, B](self: F[A, B]) {
+    def lmap[C](f: C => A): F[C, B] = macro meta.Ops.f_1
+    def rmap[C](f: B => C): F[A, C] = macro meta.Ops.f_1
+    def dimap[C, D](f: C => A)(g: B => D): F[C, D] = macro meta.Ops.f_1_1
+  }
 }
